src: mavlink: Move sys_info to its own file
