+++ src/mavlink/mod.rs
@@ -1,3 +1,4 @@
 pub mod manager;
 pub mod mavlink_camera;
+mod sys_info;
 mod utils;
